+++ app/src/main/java/ru/maksewsha/absmovies/domain/usecases/GetByKinopoiskIDCase.kt
@@ -4,7 +4,7 @@ import ru.maksewsha.absmovies.domain.models.FilmDomain
 import ru.maksewsha.absmovies.domain.repos.FilmsRepository
 
 class GetByKinopoiskIDCase(private val filmsRepository: FilmsRepository) {
-    suspend fun getByKinopoiskIDCase(id: Int): FilmDomain{
+    suspend fun getByKinopoiskIDCase(id: Long): FilmDomain{
         return filmsRepository.getByKinopoiskID(id)
     }
 }
